// types.ts
export interface Employee {
    employeeId: string;
    firstName: string;
    lastName: string;
    jobTitle: string;
    prefix: string;
    suffix: string;
    jobArea: string;
    phone: string;
    email: string;
    salary: string;
    dateJoined: string;
    dateRelieved: string;
    managerName: string;
    department: string;
    productName: string;
    city: string;
}